Big refactor for validator: splitted into several files
