+++ .vscode/settings.json
@@ -6,4 +6,5 @@
   "mochaExplorer.nodePath": "default",
   "typescript.tsdk": "node_modules/typescript/lib",
   "typescript.preferences.quoteStyle": "single",
+  "typescript.format.semicolons": "remove",
 }
