+++ src/interpreter/runtimeModel.ts
@@ -5,7 +5,7 @@ import { getPotentiallyUninitializedLazy } from '../decorators'
 import { get, is, last, List, match, raise, when } from '../extensions'
 import { Assignment, Body, Catch, Class, Describe, Entity, Environment, Expression, Field, Id, If, Literal, LiteralValue, Method, Module, Name, New, Node, Package, Program, Reference, Return, Self, Send, Singleton, Super, Test, Throw, Try, Variable } from '../model'
 import { Interpreter } from './interpreter'
-import { getUninitializedAttributesForInstantiation, loopInAssignment } from '../validator'
+import { getUninitializedAttributesForInstantiation, loopInAssignment } from '../validator/helpers'
 
 const { isArray } = Array
 const { keys, entries } = Object
