+++ src/validator/values.ts
@@ -1,0 +1,3 @@
+import { Node } from '../model'
+
+export const valuesForNodeName = (node: Node & { name?: string }): string[] => [node.name ?? '']
